requirements.txt
beautifulsoup4>=4.12,<5
requests>=2.31,<3
Pillow>=10.0,<12
